🔧 Make `template` the default `hatch` environment
Since this is a template repository:

* it seems natural to make scripts related to the template the default.
* these will be the scripts we need to run most often.

So here we make the `template` environment the default, so we can simply run:

  hatch run install

to test the installation of a generated template.

In this case, we also need to make the `docs` environment "detached", so it doesn't
inherit the scripts from the default environment.