✨ Template: Add GitHub action for deploying docs on Github
Add a GitHub action to the template for both documentation tools to deploy the
documentation to GitHub pages. This will run for any pushes on the `main` branch.

The MkDocs action uses the `hatch docs:deploy` script, but the MyST deployment is more
convoluted. For now we simply add an action based generated by the
`myst init --gh-pages` command.