👌 Template: Consistently use `docs:serve`
The MyST and MkDocs Hatch scripts for serving the documentation locally were different,
to match the corresponding commands of the tools:

* MyST: `hatch run docs:start` ~ `myst start`
* MkDocs: `hatch run docs:serve` ~ `mkdocs serve`

However, it's probably better to be uniform in our hatch script names, to avoid mixing
them up when working across packages which use both documentation tools. We pick
`serve` here, since that's what we are doing.