✨ Template: add rules to Ruff linting ignore
Ruff, the tool Hatch uses under the hood, has an extensive list of rules:

https://docs.astral.sh/ruff/rules/

Although most of them have a reasonable motivation, we might not agree with all of them
for certain reasons. For example:

https://docs.astral.sh/ruff/rules/magic-value-comparison/

would not permit comparing a certain input with a fixed value without first declaring
a constant variable. That would not be sustainable in our code, leading to a lot of
constant variable and making the code harder to understand.

Here we add a set of rules to the Ruff ignore list, and document the motivation in the
template developer docs.