🔧 Fix `hatch run check` script
The `check` script in the default environment checks if the package generated from the
template properly passes formatting and linting. However, the current command doesn't
use the hatch configuration of the generated package.

Instead of supplying the package directory as an argument, we adapt the script to first
`cd` into the directory.