🔧 Template: Add basic `tests`
Add a minimal `tests` folder with a small test that relies on a simple fixture defined
in the `conftest.py`.

Since a user of the template package typically doesn't want to keep these files when
updating their template, we exclude the `tests` directory when the `update` operation
is used.

We also set two Ruff codes to ignore when linting the package for the modules in the
`tests` directory:

* `INP001`/`implicit-namespace-package`: We don't need `__init__.py` files in the tests.
* `S101`/`assert`: Assert statements are fine in tests.