📚 Add development standards to Design notes
For now we'll add this to the template package documentation, to make it easy to share
with collaborators. It might also be desirable to place this in the template docs, but
it could also make those too verbose. Perhaps adding a link is the better option once
we converge on the developments standards here.