📚 Design: explain priorities and bump dependency one
Sometimes there might be multiple types to choose from. At the moment I'm not sure if
it's best to stick with one type specifier or specify multiple (e.g. new feature that
needs a dependency). I want to be able to identify commits that change dependencies
easily, but splitting them up makes the commit confusing, and using only the dependency
type is a bit strange for a commit that adds a new feature.