👌 Template: diversify tool options for developers
The template uses Hatch for its build backend and development environments. However, not
all developers are familiar with Hatch and/or would like to use it. We now effectively
force them to do so, or deal with manually installing the dependencies required to e.g.
run the tests.

Here we diversify options for developers:

1. We add extras for the various developer operations: `pre-commit`, `tests` and `docs`.
2. We add the `dev` dependency group, which `uv` uses by default in its environments.

See the following page for more information on dependency groups:

https://packaging.python.org/en/latest/specifications/dependency-groups/

The changes here allow users of all three approaches to contribute to a template
package.