📚 Template: update developer instructions
Now that we offer more options with regard to devops tools, we document the various
approaches in separate tabs.

To support tabs, we add the `pymdownx.superfences` and `pymdownx.tabbed` extensions
to the MkDocs config. Although `pymdownx.blocks.tab` is newer, it seems to be not fully
supported yet, at least in Material for MkDocs.

Because the MyST tab syntax does not allow indentation, and the MkDocs one with
`pymdownx.tabbed` requires it, there is quite a bit of duplication in the Jinja
template for the developer docs. We should look into approaches to avoid duplication
in the future.