📚 Design: improve discussion on type emojis
In the commit message template, add the `TYPE-` prefix to `<EMOJI>` in the subject line,
to make it more clear that this emoji has to represent the type of change.

Adapt the "important" note on not using emojis in the commit body to forbid using emojis
further down the subject line instead. Although I'm unsure if emojis in the body are a
good idea, they might have use cases, and finding the type of change is still very easy
using the `git log --oneline` option. The corresponding text in the "common concerns"
section is also adapted.