📚 Standards: move "development standards" to separate page
In hindsight, "development standards" are not really related to the design documents of
the Python `copier` template package. It might even be more sensible to add them to the
documentation of the template itself. But we leave that for when it is more complete and
polished.

Besides not being appropriate for the design section, giving the development standard
its own pages means we have to use one less header depth, which leads to more visible
header fonts and a less nested table of contents tree.

Since we now have 4 pages, it's time to add a `nav` section to the MkDocs configuration.